+++ backend/siarnaq/api/user/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from siarnaq.api.accounts.models import User
+from siarnaq.api.user.models import User
 
 # Register your models here.
 admin.site.register(User)
