+++ backend/siarnaq/api/user/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AccountsConfig(AppConfig):
+class UserConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "siarnaq.api.accounts"
+    name = "siarnaq.api.user"
