+++ backend/siarnaq/api/user/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from siarnaq.api.accounts.models import User, UserProfile
+from siarnaq.api.user.models import User, UserProfile
 
 
 class UserSerializer(serializers.ModelSerializer):
